+++ cmd/fablab/main.go
@@ -18,23 +18,24 @@ package main
 
 import (
 	"github.com/michaelquigley/pfxlog"
-	"github.com/openziti/foundation/transport"
-	"github.com/openziti/foundation/transport/quic"
-	"github.com/openziti/foundation/transport/tcp"
-	"github.com/openziti/foundation/transport/tls"
 	"github.com/openziti/fablab/cmd/fablab/subcmd"
 	"github.com/openziti/fablab/kernel/model"
 	"github.com/openziti/fablab/zitilib"
 	_ "github.com/openziti/fablab/zitilib"
 	_ "github.com/openziti/fablab/zitilib/models/characterization"
+	_ "github.com/openziti/fablab/zitilib/models/development/dilithium"
 	_ "github.com/openziti/fablab/zitilib/models/examples"
 	_ "github.com/openziti/fablab/zitilib/models/mattermozt"
+	"github.com/openziti/foundation/transport"
+	"github.com/openziti/foundation/transport/quic"
+	"github.com/openziti/foundation/transport/tcp"
+	"github.com/openziti/foundation/transport/tls"
 	"github.com/sirupsen/logrus"
 )
 
 func init() {
 	pfxlog.Global(logrus.InfoLevel)
-	pfxlog.SetPrefix("github.com/netfoundry/")
+	pfxlog.SetPrefix("github.com/openziti/")
 	transport.AddAddressParser(quic.AddressParser{})
 	transport.AddAddressParser(tls.AddressParser{})
 	transport.AddAddressParser(tcp.AddressParser{})
@@ -43,6 +44,6 @@ func init() {
 
 func main() {
 	if err := subcmd.Execute(); err != nil {
-		logrus.Fatalf("failure (%s)", err)
+		logrus.Fatalf("failure (%v)", err)
 	}
 }
