+++ zitilib/models/development/dilithium/model.go
@@ -14,12 +14,12 @@
 	limitations under the License.
 */
 
-package transwarp
+package dilithium
 
 import "github.com/openziti/fablab/kernel/model"
 
 func init() {
-	model.RegisterModel("zitilib/development/transwarp", transwarp)
+	model.RegisterModel("zitilib/development/dilithium", transwarp)
 }
 
 var transwarp = &model.Model{
